+++ src/components/data-display/Avatar/Avatar.tsx
@@ -1,6 +1,7 @@
 import { Avatar as AntAvatar } from 'antd'
 import { type AvatarProps as AntAvatarProps } from 'antd'
 import { ConfigProvider } from 'src/components'
+import './avatar.css'
 
 export interface IAvatarProps extends AntAvatarProps {}