+++ src/components/data-display/Popover/Popover.stories.tsx
@@ -1,6 +1,9 @@
 import { Popover, type IPopoverProps } from 'src/components/data-display/Popover/Popover'
 import { type Meta, type StoryObj } from '@storybook/react'
 import { Button } from 'src/components/general/Button/Button'
+import { Tag } from 'src/components/data-display/Tag/Tag'
+import { Typography } from 'src/components/general/Typography/Typography'
+import { Flex } from 'src/components/layout/Flex/Flex'
 
 const meta: Meta<typeof Popover> = {
   title: 'Components/Data Display/Popover',
@@ -37,4 +40,33 @@ export const Primary: Story = {
     },
   },
   render: PrimaryTemplate,
+}
+
+const PopoverWithLinksTemplate = (args: IPopoverProps) => {
+  const content = (
+    <Flex gap="small" vertical>
+      <Typography.Link href="#link1">Link Example 1</Typography.Link>
+      <Typography.Link href="#link2">Link Example 2</Typography.Link>
+      <Typography.Link href="#link3">Link Example 3</Typography.Link>
+    </Flex>
+  )
+
+  return (
+    <>
+      <Popover content={content} title="Details" trigger={args.trigger}>
+        <Tag style={{ margin: '0', cursor: 'pointer' }}>{args.children ?? '3'}</Tag>
+      </Popover>
+    </>
+  )
+}
+
+export const ExampleWithLinks: Story = {
+  args: { trigger: 'click', children: '3' },
+  argTypes: {
+    trigger: {
+      control: 'select',
+      options: ['click', 'hover', 'focus'],
+    },
+  },
+  render: PopoverWithLinksTemplate,
 }