+++ src/components/data-entry/Checkbox/Checkbox.tsx
@@ -1,11 +1,13 @@
 import { Checkbox as AntCheckbox } from 'antd'
 import { type CheckboxProps as AntCheckboxProps } from 'antd'
+import { type CheckboxGroupProps as AntCheckboxGroupProps } from 'antd/es/checkbox/Group'
 import { ConfigProvider } from 'src/components'
 import './checkbox.css'
 
 export interface ICheckboxProps extends AntCheckboxProps {}
+export interface ICheckboxGroupProps extends AntCheckboxGroupProps {}
 
-export const Checkbox = (props: ICheckboxProps) => {
+const CheckboxComponent = (props: ICheckboxProps) => {
   return (
     <ConfigProvider>
       <AntCheckbox {...props} />
@@ -13,4 +15,12 @@ export const Checkbox = (props: ICheckboxProps) => {
   )
 }
 
-Checkbox.Group = AntCheckbox.Group
+const CheckboxGroup = (props: ICheckboxGroupProps) => {
+  return (
+    <ConfigProvider>
+      <AntCheckbox.Group {...props} />
+    </ConfigProvider>
+  )
+}
+
+export const Checkbox = Object.assign(CheckboxComponent, { Group: CheckboxGroup })