+++ src/components/general/Button/Button.tsx
@@ -7,20 +7,14 @@ import { ConfigProvider } from 'src/components/other/ConfigProvider/ConfigProvid
 import { type ReactNode } from 'react'
 import './button.css'
 
-export interface IButtonProps extends Omit<AntButtonProps, 'variant'> {
-  /**
-   * @deprecated This variant is a temporary fix for new icons.
-   * Use this variant only with new icons to align the icon and text centered.
-   * This will be removed once all icons are updated.
-   */
-  variant?: 'with-new-icon'
+export interface IButtonProps extends AntButtonProps {
   icon?: ReactNode
+  /**
+   * Text color inheritance from parent element.
+   */
+  textColor?: 'inherit'
 }
 export const Button = (props: IButtonProps) => {
-  const classMap = {
-    'with-new-icon': 'u-display-flex u-align-items-center u-justify-center',
-  }
-
   const buttonIcon =
     React.isValidElement<IIconProps>(props.icon) && props.icon.type === Icon ? (
       <Icon {...props.icon.props} color={props.icon.props.color ?? 'inherit'} />
@@ -28,13 +22,18 @@ export const Button = (props: IButtonProps) => {
       props.icon
     )
 
-  const extraClass = props.variant === 'with-new-icon' ? classMap['with-new-icon'] : ''
+  const { variant, textColor, style, ...restProps } = props
 
-  const { variant, ...restProps } = props
+  const buttonStyle = textColor === 'inherit' ? { ...style, color: 'inherit' } : style
 
   return (
     <ConfigProvider>
-      <AntButton {...restProps} icon={buttonIcon} className={`${props.className ?? ''} ${extraClass}`}>
+      <AntButton
+        {...restProps}
+        icon={buttonIcon}
+        style={buttonStyle}
+        variant={variant}
+        className={`${props.className ?? ''}`}>
         {props.children}
       </AntButton>
     </ConfigProvider>